requirements.txt
azure_storage==0.37.0
matplotlib==3.5.2
numpy==1.22.3
pandas==1.5.1
python-dotenv==0.21.0
